refactor: extract backend routing and service seams
