+++ codoxear/http/routes/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
