+++ codoxear/server.py
@@ -65,6 +65,16 @@ from .util import read_jsonl_from_offset as _read_jsonl_from_offset_impl
 from .util import read_session_meta_payload as _read_session_meta_payload_impl
 from .util import subagent_parent_thread_id as _subagent_parent_thread_id
 from .voice_push import VoicePushCoordinator
+from .http.routes import assets as _http_assets_routes
+from .http.routes import auth as _http_auth_routes
+from .http.routes import files as _http_file_routes
+from .http.routes import notifications as _http_notification_routes
+from .http.routes import sessions_read as _http_session_read_routes
+from .http.routes import sessions_write as _http_session_write_routes
+from .sessions import live_payloads as _session_live_payloads
+from .sessions import payloads as _session_payloads
+from .sessions.sidebar_state import SidebarStateFacade
+from .pi import ui_bridge as _pi_ui_bridge
 
 
 LOG = logging.getLogger(__name__)
@@ -1937,13 +1947,7 @@ def _listed_session_row(manager: "SessionManager", session_id: str) -> dict[str,
 def _session_details_payload(
     manager: "SessionManager", session_id: str
 ) -> dict[str, Any]:
-    row = _listed_session_row(manager, session_id)
-    if row is not None:
-        return {"ok": True, "session": _normalize_session_cwd_row(row)}
-    historical_row = _historical_session_row(session_id)
-    if historical_row is not None:
-        return {"ok": True, "session": _normalize_session_cwd_row(dict(historical_row))}
-    raise KeyError("unknown session")
+    return _session_payloads.session_details_payload(manager, session_id)
 
 
 def _clean_recent_cwd(value: Any) -> str | None:
@@ -3871,138 +3875,19 @@ def _resolved_session_run_settings(
 def _session_context_usage_payload(
     s: Session, token_val: dict[str, Any] | None
 ) -> dict[str, Any] | None:
-    if s.backend != "pi":
-        return None
-    context_window = None
-    model_provider, _preferred_auth_method, model, _reasoning_effort = (
-        _resolved_session_run_settings(s)
-    )
-    if model_provider is not None and model is not None:
-        context_window = _pi_model_context_window(model_provider, model)
-    if (not isinstance(context_window, int) or context_window <= 0) and isinstance(
-        token_val, dict
-    ):
-        token_context_window = token_val.get("context_window")
-        if isinstance(token_context_window, int) and token_context_window > 0:
-            context_window = token_context_window
-    if not isinstance(context_window, int) or context_window <= 0:
-        return None
-    used_tokens = 0
-    if isinstance(token_val, dict):
-        raw_used_tokens = token_val.get("tokens_in_context")
-        if isinstance(raw_used_tokens, int) and raw_used_tokens > 0:
-            used_tokens = raw_used_tokens
-    used_tokens = min(max(used_tokens, 0), context_window)
-    percent_used = int(round((used_tokens / context_window) * 100.0)) if context_window > 0 else 0
-    return {
-        "used_tokens": used_tokens,
-        "total_tokens": context_window,
-        "percent_used": percent_used,
-    }
+    return _session_payloads.session_context_usage_payload(s, token_val)
 
 
 def _session_diagnostics_payload(
     manager: "SessionManager", session_id: str, s: Session, state: dict[str, Any]
 ) -> dict[str, Any]:
-    state = _validated_session_state(state)
-    token_val: dict[str, Any] | None = None
-    st_token = state.get("token")
-    if isinstance(st_token, dict) or st_token is None:
-        token_val = (
-            st_token
-            if isinstance(st_token, dict)
-            else (s.token if isinstance(s.token, dict) else None)
-        )
-    model_provider, preferred_auth_method, model, reasoning_effort = _resolved_session_run_settings(s)
-    service_tier = s.service_tier
-    sidebar_meta = manager.sidebar_meta_get(session_id)
-    cwd_path = _safe_expanduser(Path(s.cwd))
-    if not cwd_path.is_absolute():
-        cwd_path = cwd_path.resolve()
-    git_branch = _current_git_branch(cwd_path)
-    updated_ts = _display_updated_ts(s)
-    elapsed_s = max(0.0, time.time() - updated_ts)
-    time_priority = _priority_from_elapsed_seconds(elapsed_s)
-    base_priority = _clip01(time_priority + float(sidebar_meta["priority_offset"]))
-    blocked = sidebar_meta["dependency_session_id"] is not None
-    snoozed = (
-        sidebar_meta["snooze_until"] is not None
-        and float(sidebar_meta["snooze_until"]) > time.time()
-    )
-    final_priority = 0.0 if (snoozed or blocked) else base_priority
-    busy, broker_busy = _display_session_busy(manager, session_id, s, state)
-    durable_session_id = _durable_session_id_for_live_session(s)
-    return {
-        "session_id": durable_session_id,
-        "runtime_id": s.session_id,
-        "thread_id": s.thread_id,
-        "agent_backend": s.agent_backend,
-        "backend": s.backend,
-        "owned": bool(s.owned),
-        "transport": s.transport,
-        "cwd": s.cwd,
-        "start_ts": float(s.start_ts),
-        "updated_ts": updated_ts,
-        "log_path": str(s.log_path) if s.log_path is not None else None,
-        "session_file_path": _display_source_path(s),
-        "broker_pid": int(s.broker_pid),
-        "codex_pid": int(s.codex_pid),
-        "busy": bool(busy),
-        "broker_busy": broker_busy,
-        "queue_len": manager._queue_len(session_id),
-        "token": token_val,
-        "context_usage": _session_context_usage_payload(s, token_val),
-        "model_provider": model_provider,
-        "preferred_auth_method": preferred_auth_method,
-        "provider_choice": _provider_choice_for_backend(
-            backend=s.backend,
-            model_provider=model_provider,
-            preferred_auth_method=preferred_auth_method,
-        ),
-        "model": model,
-        "reasoning_effort": reasoning_effort,
-        "service_tier": service_tier,
-        "tmux_session": s.tmux_session,
-        "tmux_window": s.tmux_window,
-        "git_branch": git_branch,
-        "time_priority": time_priority,
-        "base_priority": base_priority,
-        "final_priority": final_priority,
-        "priority_offset": sidebar_meta["priority_offset"],
-        "snooze_until": sidebar_meta["snooze_until"],
-        "dependency_session_id": sidebar_meta["dependency_session_id"],
-        "todo_snapshot": _todo_snapshot_payload_for_session(s),
-    }
+    return _session_payloads.session_diagnostics_payload(manager, session_id, s, state)
 
 
 def _session_workspace_payload(
     manager: "SessionManager", session_id: str
 ) -> dict[str, Any]:
-    manager.refresh_session_meta(session_id, strict=False)
-    s = manager.get_session(session_id)
-    if not s:
-        historical_row = _historical_session_row(session_id)
-        if historical_row is None:
-            historical_row = _listed_session_row(manager, session_id)
-        if historical_row is None:
-            raise KeyError("unknown session")
-        return {
-            "ok": True,
-            "session_id": str(historical_row.get("session_id") or session_id),
-            "runtime_id": None,
-            "diagnostics": None,
-            "queue": {"items": []},
-        }
-    diagnostics = _session_diagnostics_payload(
-        manager, session_id, s, manager.get_state(session_id)
-    )
-    return {
-        "ok": True,
-        "session_id": _durable_session_id_for_live_session(s),
-        "runtime_id": s.session_id,
-        "diagnostics": diagnostics,
-        "queue": {"items": manager.queue_list(session_id)},
-    }
+    return _session_payloads.session_workspace_payload(manager, session_id)
 
 
 def _session_live_payload(
@@ -4014,166 +3899,26 @@ def _session_live_payload(
     bridge_offset: int = 0,
     requests_version: str | None = None,
 ) -> dict[str, Any]:
-    manager.refresh_session_meta(session_id, strict=False)
-    s = manager.get_session(session_id)
-    if not s:
-        historical_row = _historical_session_row(session_id)
-        if historical_row is None:
-            historical_row = _listed_session_row(manager, session_id)
-        if historical_row is None:
-            raise KeyError("unknown session")
-        try:
-            page = manager.get_messages_page(
-                session_id, offset=max(0, int(offset)), init=(offset <= 0), limit=80, before=0
-            )
-        except KeyError:
-            page = {"events": [], "offset": max(0, int(offset)), "has_older": False, "next_before": 0}
-        durable_session_id = str(historical_row.get("session_id") or session_id)
-        bridge_session_key = _clean_optional_text(
-            historical_row.get("resume_session_id")
-        ) or durable_session_id
-        bridge_events, next_bridge_offset = manager._bridge_events_since(
-            bridge_session_key,
-            offset=bridge_offset,
-        )
-        base_events = page.get("events") if isinstance(page.get("events"), list) else []
-        return {
-            "ok": True,
-            "session_id": durable_session_id,
-            "runtime_id": None,
-            "offset": int(page.get("offset", max(0, int(offset))) or 0),
-            "live_offset": 0,
-            "bridge_offset": next_bridge_offset,
-            "has_older": bool(page.get("has_older")),
-            "next_before": int(page.get("next_before", 0) or 0),
-            "busy": False,
-            "events": [*base_events, *bridge_events],
-            "requests_version": "",
-            "requests": [],
-            "token": None,
-            "context_usage": None,
-        }
-    page = manager.get_messages_page(
-        session_id, offset=max(0, int(offset)), init=(offset <= 0), limit=80, before=0
+    return _session_live_payloads.session_live_payload(
+        manager,
+        session_id,
+        offset=offset,
+        live_offset=live_offset,
+        bridge_offset=bridge_offset,
+        requests_version=requests_version,
     )
-    state = manager.get_state(session_id)
-    busy, _broker_busy = _display_session_busy(manager, session_id, s, state)
-    state_token = state.get("token") if isinstance(state, dict) else None
-    token_val = (
-        state_token
-        if isinstance(state_token, dict)
-        else (s.token if isinstance(s.token, dict) else None)
-    )
-    requests: list[dict[str, Any]] = []
-    if s.backend == "pi":
-        requests_payload = manager.get_ui_state(session_id)
-        live_requests = requests_payload.get("requests")
-        if isinstance(live_requests, list):
-            requests = [item for item in live_requests if isinstance(item, dict)]
-    current_requests_version = _ui_requests_version(requests)
-    events = page.get("events")
-    merged_events = events if isinstance(events, list) else []
-    next_live_offset = max(0, int(live_offset))
-    next_bridge_offset = max(0, int(bridge_offset))
-    if _session_supports_live_pi_ui(s):
-        streamed_payload = _pi_live_messages_payload(manager, s, offset=live_offset)
-        next_live_offset = int(
-            streamed_payload.get("offset", max(0, int(live_offset)))
-            or max(0, int(live_offset))
-        )
-        merged_events = _merge_pi_live_message_events(
-            merged_events,
-            [
-                item
-                for item in (streamed_payload.get("events") or [])
-                if isinstance(item, dict)
-            ],
-        )
-    bridge_events, next_bridge_offset = manager._bridge_events_since(
-        _durable_session_id_for_live_session(s),
-        offset=bridge_offset,
-    )
-    if bridge_events:
-        merged_events = [*merged_events, *bridge_events]
-    payload: dict[str, Any] = {
-        "ok": True,
-        "session_id": _durable_session_id_for_live_session(s),
-        "runtime_id": s.session_id,
-        "offset": int(page.get("offset", max(0, int(offset))) or 0),
-        "live_offset": next_live_offset,
-        "bridge_offset": next_bridge_offset,
-        "has_older": bool(page.get("has_older")),
-        "next_before": int(page.get("next_before", 0) or 0),
-        "busy": bool(busy),
-        "events": merged_events,
-        "requests_version": current_requests_version,
-        "token": token_val,
-        "context_usage": _session_context_usage_payload(s, token_val),
-        "transport_state": s.bridge_transport_state,
-        "transport_error": s.bridge_transport_error,
-    }
-    if requests_version != current_requests_version:
-        payload["requests"] = requests
-    return payload
 
 
 def _pi_live_messages_payload(
     manager: "SessionManager", session: Session, *, offset: int = 0
 ) -> dict[str, Any]:
-    if not _session_supports_live_pi_ui(session):
-        return {"offset": max(0, int(offset)), "events": []}
-    try:
-        payload = manager._sock_call(
-            session.sock_path,
-            {"cmd": "live_messages", "offset": max(0, int(offset))},
-            timeout_s=1.5,
-        )
-    except Exception:
-        return {"offset": max(0, int(offset)), "events": []}
-    events = payload.get("events")
-    return {
-        "offset": int(payload.get("offset", max(0, int(offset))) or 0),
-        "events": [item for item in events if isinstance(item, dict)]
-        if isinstance(events, list)
-        else [],
-    }
+    return _session_live_payloads.pi_live_messages_payload(manager, session, offset=offset)
 
 
 def _merge_pi_live_message_events(
     durable_events: list[dict[str, Any]], streamed_events: list[dict[str, Any]]
 ) -> list[dict[str, Any]]:
-    merged = list(durable_events)
-    durable_turn_ids = {
-        str(event.get("turn_id"))
-        for event in durable_events
-        if event.get("role") == "assistant"
-        and isinstance(event.get("turn_id"), str)
-        and str(event.get("turn_id") or "")
-    }
-    tail_durable_text = ""
-    for event in reversed(durable_events):
-        if event.get("role") != "assistant":
-            if event.get("role") == "user":
-                break
-            continue
-        text = event.get("text")
-        if isinstance(text, str) and text.strip():
-            tail_durable_text = text.strip()
-            break
-    for event in streamed_events:
-        if event.get("role") != "assistant":
-            merged.append(event)
-            continue
-        turn_id = (
-            event.get("turn_id") if isinstance(event.get("turn_id"), str) else None
-        )
-        text = str(event.get("text") or "").strip()
-        if turn_id and turn_id in durable_turn_ids:
-            continue
-        if bool(event.get("completed")) and text and text == tail_durable_text:
-            continue
-        merged.append(event)
-    return merged
+    return _session_live_payloads.merge_pi_live_message_events(durable_events, streamed_events)
 
 
 def _supports_web_control(meta: dict[str, Any]) -> bool:
@@ -4207,6 +3952,7 @@ class SessionManager:
         self._recent_cwds: dict[str, float] = {}
         self._cwd_groups: dict[str, dict[str, Any]] = {}
         self._prune_missing_workspace_dirs = True
+        self._sidebar_state = SidebarStateFacade(self)
         self._harness_last_injected: dict[str, float] = {}
         self._harness_last_injected_scope: dict[str, float] = {}
         self._load_harness()
@@ -4240,6 +3986,14 @@ class SessionManager:
             target=self._voice_push_scan_loop, name="voice-push-scan", daemon=True
         )
         self._voice_push_scan_thr.start()
+
+    def _sidebar_state_facade(self) -> SidebarStateFacade:
+        facade = getattr(self, "_sidebar_state", None)
+        if isinstance(facade, SidebarStateFacade):
+            return facade
+        facade = SidebarStateFacade(self)
+        self._sidebar_state = facade
+        return facade
 
     def stop(self) -> None:
         self._stop.set()
@@ -4645,26 +4399,7 @@ class SessionManager:
             self._clear_deleted_session_state(durable_session_id)
 
     def _persist_session_ui_state(self) -> None:
-        db = getattr(self, "_page_state_db", None)
-        if db is None:
-            return
-        with self._lock:
-            aliases_src = dict(self._aliases)
-            sidebar_src = dict(self._sidebar_meta)
-            hidden_keys = set(self._hidden_sessions)
-        aliases: dict[SessionRef, str] = {}
-        for key, value in aliases_src.items():
-            ref = key if isinstance(key, tuple) and len(key) == 2 else self._page_state_ref_for_session_id(str(key))
-            if ref is None or not isinstance(value, str) or not value.strip():
-                continue
-            aliases[ref] = value
-        sidebar_meta: dict[SessionRef, dict[str, Any]] = {}
-        for key, value in sidebar_src.items():
-            ref = key if isinstance(key, tuple) and len(key) == 2 else self._page_state_ref_for_session_id(str(key))
-            if ref is None or not isinstance(value, dict):
-                continue
-            sidebar_meta[ref] = dict(value)
-        db.save_session_ui_state(aliases, sidebar_meta, hidden_keys)
+        self._sidebar_state_facade().persist_session_ui_state()
 
     def _persist_files(self) -> None:
         db = getattr(self, "_page_state_db", None)
@@ -4931,25 +4666,19 @@ class SessionManager:
         os.replace(tmp, HARNESS_PATH)
 
     def _load_aliases(self) -> None:
-        aliases, _sidebar_meta, _hidden_keys = self._page_state_db.load_session_ui_state()
-        with self._lock:
-            self._aliases = aliases
+        self._sidebar_state_facade().load_aliases()
 
     def _save_aliases(self) -> None:
         self._persist_session_ui_state()
 
     def _load_sidebar_meta(self) -> None:
-        _aliases, sidebar_meta, _hidden_keys = self._page_state_db.load_session_ui_state()
-        with self._lock:
-            self._sidebar_meta = sidebar_meta
+        self._sidebar_state_facade().load_sidebar_meta()
 
     def _save_sidebar_meta(self) -> None:
         self._persist_session_ui_state()
 
     def _load_hidden_sessions(self) -> None:
-        _aliases, _sidebar_meta, hidden_keys = self._page_state_db.load_session_ui_state()
-        with self._lock:
-            self._hidden_sessions = hidden_keys
+        self._sidebar_state_facade().load_hidden_sessions()
 
     def _save_hidden_sessions(self) -> None:
         self._persist_session_ui_state()
@@ -4961,20 +4690,9 @@ class SessionManager:
         resume_session_id: str | None,
         backend: str | None,
     ) -> set[str]:
-        keys: set[str] = set()
-        session_clean = _clean_optional_text(session_id)
-        thread_clean = _clean_optional_text(thread_id)
-        resume_clean = _clean_optional_text(resume_session_id)
-        backend_clean = normalize_agent_backend(backend, default="codex")
-        if session_clean:
-            keys.add(session_clean)
-            keys.add(f"session:{session_clean}")
-        if thread_clean:
-            keys.add(f"thread:{backend_clean}:{thread_clean}")
-        if resume_clean:
-            keys.add(f"resume:{backend_clean}:{resume_clean}")
-            keys.add(_historical_session_id(backend_clean, resume_clean))
-        return keys
+        return self._sidebar_state_facade().hidden_session_keys(
+            session_id, thread_id, resume_session_id, backend
+        )
 
     def _session_is_hidden(
         self,
@@ -4983,25 +4701,12 @@ class SessionManager:
         resume_session_id: str | None,
         backend: str | None,
     ) -> bool:
-        with self._lock:
-            hidden = set(getattr(self, "_hidden_sessions", set()))
-        return bool(
-            hidden
-            and hidden.intersection(
-                self._hidden_session_keys(
-                    session_id, thread_id, resume_session_id, backend
-                )
-            )
+        return self._sidebar_state_facade().session_is_hidden(
+            session_id, thread_id, resume_session_id, backend
         )
 
     def _hide_session(self, session_id: str) -> None:
-        with self._lock:
-            hidden = getattr(self, "_hidden_sessions", None)
-            if not isinstance(hidden, set):
-                self._hidden_sessions = set()
-                hidden = self._hidden_sessions
-            hidden.add(session_id)
-        self._save_hidden_sessions()
+        self._sidebar_state_facade().hide_session(session_id)
 
     def _hide_session_identity_values(
         self,
@@ -5010,37 +4715,15 @@ class SessionManager:
         resume_session_id: str | None,
         backend: str | None,
     ) -> None:
-        hidden_keys = self._hidden_session_keys(
-            session_id,
-            thread_id,
-            resume_session_id,
-            backend,
+        self._sidebar_state_facade().hide_session_identity_values(
+            session_id, thread_id, resume_session_id, backend
         )
-        with self._lock:
-            hidden = getattr(self, "_hidden_sessions", None)
-            if not isinstance(hidden, set):
-                self._hidden_sessions = set()
-                hidden = self._hidden_sessions
-            hidden.update(hidden_keys)
-        self._save_hidden_sessions()
 
     def _hide_session_identity(self, s: Session) -> None:
-        self._hide_session_identity_values(
-            s.session_id,
-            s.thread_id,
-            s.resume_session_id,
-            s.agent_backend or s.backend,
-        )
+        self._sidebar_state_facade().hide_session_identity(s)
 
     def _unhide_session(self, session_id: str) -> None:
-        changed = False
-        with self._lock:
-            hidden = getattr(self, "_hidden_sessions", None)
-            if isinstance(hidden, set) and session_id in hidden:
-                hidden.remove(session_id)
-                changed = True
-        if changed:
-            self._save_hidden_sessions()
+        self._sidebar_state_facade().unhide_session(session_id)
 
     def set_created_session_name(
         self,
@@ -5050,102 +4733,24 @@ class SessionManager:
         backend: Any = None,
         name: Any,
     ) -> str:
-        alias = _clean_alias(name)
-        if not alias:
-            return ""
-        session_id_clean = _clean_optional_text(session_id)
-        runtime_id_clean = _clean_optional_text(runtime_id)
-        ref = None
-        if session_id_clean is not None:
-            ref = self._page_state_ref_for_session_id(session_id_clean)
-        if ref is None and runtime_id_clean is not None:
-            ref = self._page_state_ref_for_session_id(runtime_id_clean)
-        backend_clean = (
-            normalize_agent_backend(backend, default="codex")
-            if backend is not None
-            else None
+        return self._sidebar_state_facade().set_created_session_name(
+            session_id=session_id,
+            runtime_id=runtime_id,
+            backend=backend,
+            name=name,
         )
-        target: SessionRef | str | None = ref
-        if target is None and session_id_clean is not None and backend_clean is not None:
-            target = (backend_clean, session_id_clean)
-        if target is None and runtime_id_clean is not None and backend_clean is not None:
-            target = (backend_clean, runtime_id_clean)
-        if target is None:
-            raise KeyError("unknown session")
-        with self._lock:
-            self._aliases[target] = alias
-            if session_id_clean is not None:
-                self._aliases.pop(session_id_clean, None)
-            if runtime_id_clean is not None:
-                self._aliases.pop(runtime_id_clean, None)
-        save_aliases = getattr(self, "_save_aliases", None)
-        if callable(save_aliases):
-            save_aliases()
-        return alias
 
     def alias_set(self, session_id: str, name: str) -> str:
-        alias = _clean_alias(name)
-        ref = self._page_state_ref_for_session_id(session_id)
-        runtime_id = self._runtime_session_id_for_identifier(session_id)
-        if ref is None:
-            raise KeyError("unknown session")
-        with self._lock:
-            if alias:
-                self._aliases[ref] = alias
-            else:
-                if runtime_id is not None:
-                    self._aliases.pop(runtime_id, None)
-                self._aliases.pop(ref, None)
-        self._save_aliases()
-        return alias
+        return self._sidebar_state_facade().alias_set(session_id, name)
 
     def alias_get(self, session_id: str) -> str:
-        ref = self._page_state_ref_for_session_id(session_id)
-        if ref is None:
-            return ""
-        with self._lock:
-            alias = self._aliases.get(ref)
-            if alias is None:
-                alias = self._aliases.get(session_id)
-        return alias if isinstance(alias, str) else ""
+        return self._sidebar_state_facade().alias_get(session_id)
 
     def alias_clear(self, session_id: str) -> None:
-        ref = self._page_state_ref_for_session_id(session_id)
-        if ref is None:
-            return
-        with self._lock:
-            if ref not in self._aliases and session_id not in self._aliases:
-                return
-            self._aliases.pop(ref, None)
-            self._aliases.pop(session_id, None)
-        self._save_aliases()
+        self._sidebar_state_facade().alias_clear(session_id)
 
     def sidebar_meta_get(self, session_id: str) -> dict[str, Any]:
-        ref = self._page_state_ref_for_session_id(session_id)
-        if ref is None:
-            raise KeyError("unknown session")
-        with self._lock:
-            meta_map = getattr(self, "_sidebar_meta", None)
-            entry = None
-            if isinstance(meta_map, dict):
-                entry = meta_map.get(ref)
-                if entry is None:
-                    entry = meta_map.get(session_id)
-        if not isinstance(entry, dict):
-            return {
-                "priority_offset": 0.0,
-                "snooze_until": None,
-                "dependency_session_id": None,
-                "focused": False,
-            }
-        return {
-            "priority_offset": _clean_priority_offset(entry.get("priority_offset")),
-            "snooze_until": _clean_snooze_until(entry.get("snooze_until")),
-            "dependency_session_id": _clean_dependency_session_id(
-                entry.get("dependency_session_id")
-            ),
-            "focused": bool(entry.get("focused")),
-        }
+        return self._sidebar_state_facade().sidebar_meta_get(session_id)
 
     def sidebar_meta_set(
         self,
@@ -5155,90 +4760,15 @@ class SessionManager:
         snooze_until: Any,
         dependency_session_id: Any,
     ) -> dict[str, Any]:
-        offset = _clean_priority_offset(priority_offset)
-        snooze_until_clean = _clean_snooze_until(snooze_until)
-        dependency_clean = _clean_dependency_session_id(dependency_session_id)
-        ref = self._page_state_ref_for_session_id(session_id)
-        if ref is None:
-            raise KeyError("unknown session")
-        dep_ref = None
-        if dependency_clean is not None:
-            dep_ref = self._page_state_ref_for_session_id(dependency_clean)
-            if dep_ref is None:
-                raise ValueError("dependency session not found")
-            if dep_ref == ref:
-                raise ValueError("session cannot depend on itself")
-        with self._lock:
-            meta_map = getattr(self, "_sidebar_meta", None)
-            if not isinstance(meta_map, dict):
-                self._sidebar_meta = {}
-                meta_map = self._sidebar_meta
-            existing = meta_map.get(ref)
-            entry = {"priority_offset": offset}
-            if isinstance(existing, dict) and existing.get("focused"):
-                entry["focused"] = True
-            if snooze_until_clean is not None:
-                entry["snooze_until"] = snooze_until_clean
-            if dep_ref is not None:
-                entry["dependency_session_id"] = dep_ref[1]
-            meta_map[ref] = entry
-        self._save_sidebar_meta()
-        return {
-            "priority_offset": offset,
-            "snooze_until": snooze_until_clean,
-            "dependency_session_id": dep_ref[1] if dep_ref is not None else None,
-            "focused": bool(isinstance(existing, dict) and existing.get("focused")),
-        }
+        return self._sidebar_state_facade().sidebar_meta_set(
+            session_id,
+            priority_offset=priority_offset,
+            snooze_until=snooze_until,
+            dependency_session_id=dependency_session_id,
+        )
 
     def focus_set(self, session_id: str, focused: Any) -> bool:
-        focused_clean = _clean_optional_bool(focused)
-        if focused_clean is None:
-            raise ValueError("focused must be a boolean")
-        ref = self._page_state_ref_for_session_id(session_id)
-        if ref is None:
-            raise KeyError("unknown session")
-        with self._lock:
-            meta_map = getattr(self, "_sidebar_meta", None)
-            if not isinstance(meta_map, dict):
-                self._sidebar_meta = {}
-                meta_map = self._sidebar_meta
-            entry = meta_map.get(ref)
-            next_entry = dict(entry) if isinstance(entry, dict) else {}
-            if focused_clean:
-                next_entry["focused"] = True
-            else:
-                next_entry.pop("focused", None)
-            next_entry["priority_offset"] = _clean_priority_offset(
-                next_entry.get("priority_offset")
-            )
-            next_entry["snooze_until"] = _clean_snooze_until(
-                next_entry.get("snooze_until")
-            )
-            next_entry["dependency_session_id"] = _clean_dependency_session_id(
-                next_entry.get("dependency_session_id")
-            )
-            if (
-                not next_entry.get("focused")
-                and not next_entry.get("priority_offset")
-                and next_entry.get("snooze_until") is None
-                and next_entry.get("dependency_session_id") is None
-            ):
-                meta_map.pop(ref, None)
-            else:
-                cleaned_entry = {
-                    "priority_offset": float(next_entry.get("priority_offset") or 0.0)
-                }
-                if next_entry.get("focused"):
-                    cleaned_entry["focused"] = True
-                if next_entry.get("snooze_until") is not None:
-                    cleaned_entry["snooze_until"] = next_entry["snooze_until"]
-                if next_entry.get("dependency_session_id") is not None:
-                    cleaned_entry["dependency_session_id"] = next_entry[
-                        "dependency_session_id"
-                    ]
-                meta_map[ref] = cleaned_entry
-        self._save_sidebar_meta()
-        return focused_clean
+        return self._sidebar_state_facade().focus_set(session_id, focused)
 
     def edit_session(
         self,
@@ -5249,53 +4779,13 @@ class SessionManager:
         snooze_until: Any,
         dependency_session_id: Any,
     ) -> tuple[str, dict[str, Any]]:
-        alias = _clean_alias(name)
-        offset = _clean_priority_offset(priority_offset)
-        snooze_until_clean = _clean_snooze_until(snooze_until)
-        dependency_clean = _clean_dependency_session_id(dependency_session_id)
-        ref = self._page_state_ref_for_session_id(session_id)
-        runtime_id = self._runtime_session_id_for_identifier(session_id)
-        if ref is None:
-            raise KeyError("unknown session")
-        dep_ref = None
-        if dependency_clean is not None:
-            dep_ref = self._page_state_ref_for_session_id(dependency_clean)
-            if dep_ref is None:
-                raise ValueError("dependency session not found")
-            if dep_ref == ref:
-                raise ValueError("session cannot depend on itself")
-        with self._lock:
-            aliases = getattr(self, "_aliases", None)
-            if not isinstance(aliases, dict):
-                self._aliases = {}
-                aliases = self._aliases
-            if alias:
-                aliases[ref] = alias
-            else:
-                if runtime_id is not None:
-                    aliases.pop(runtime_id, None)
-                aliases.pop(ref, None)
-            meta_map = getattr(self, "_sidebar_meta", None)
-            if not isinstance(meta_map, dict):
-                self._sidebar_meta = {}
-                meta_map = self._sidebar_meta
-            existing = meta_map.get(ref)
-            entry = {"priority_offset": offset}
-            if isinstance(existing, dict) and existing.get("focused"):
-                entry["focused"] = True
-            if snooze_until_clean is not None:
-                entry["snooze_until"] = snooze_until_clean
-            if dep_ref is not None:
-                entry["dependency_session_id"] = dep_ref[1]
-            meta_map[ref] = entry
-        self._save_aliases()
-        self._save_sidebar_meta()
-        return alias, {
-            "priority_offset": offset,
-            "snooze_until": snooze_until_clean,
-            "dependency_session_id": dep_ref[1] if dep_ref is not None else None,
-            "focused": bool(isinstance(existing, dict) and existing.get("focused")),
-        }
+        return self._sidebar_state_facade().edit_session(
+            session_id,
+            name=name,
+            priority_offset=priority_offset,
+            snooze_until=snooze_until,
+            dependency_session_id=dependency_session_id,
+        )
 
     def _clear_deleted_session_state(self, session_id: str) -> None:
         changed_sidebar = False
@@ -8856,154 +8346,15 @@ class SessionManager:
         return resp
 
     def get_ui_state(self, session_id: str) -> dict[str, Any]:
-        runtime_id = self._runtime_session_id_for_identifier(session_id)
-        if runtime_id is None:
-            raise KeyError("unknown session")
-        self.refresh_session_meta(runtime_id, strict=False)
-        with self._lock:
-            s = self._sessions.get(runtime_id)
-            if not s:
-                raise KeyError("unknown session")
-            if s.backend != "pi":
-                raise ValueError("ui interactions are only supported for pi sessions")
-            sock = s.sock_path
-        try:
-            resp = self._sock_call(sock, {"cmd": "ui_state"}, timeout_s=1.5)
-        except Exception:
-            if not _pid_alive(s.broker_pid) and not _pid_alive(s.codex_pid):
-                with self._lock:
-                    self._sessions.pop(runtime_id, None)
-                self._clear_deleted_session_state(runtime_id)
-                _unlink_quiet(sock)
-                _unlink_quiet(sock.with_suffix(".json"))
-                raise KeyError("unknown session")
-            raise ValueError("broker unavailable")
-
-        if resp.get("error") == "unknown cmd":
-            if _session_supports_live_pi_ui(s):
-                raise ValueError(
-                    "live ui interactions are unavailable for this pi session"
-                )
-            # Gracefully handle older brokers that don't support ui_state yet.
-            return {"requests": []}
-
-        error = resp.get("error")
-        if isinstance(error, str) and error:
-            raise ValueError(error)
-        return _sanitize_pi_ui_state_payload(resp)
+        return _pi_ui_bridge.get_ui_state(self, session_id)
 
     def get_session_commands(self, session_id: str) -> dict[str, Any]:
-        runtime_id = self._runtime_session_id_for_identifier(session_id)
-        if runtime_id is None:
-            raise KeyError("unknown session")
-        self.refresh_session_meta(runtime_id, strict=False)
-        with self._lock:
-            s = self._sessions.get(runtime_id)
-            if not s:
-                raise KeyError("unknown session")
-            if s.backend != "pi":
-                raise ValueError("command listing is only supported for pi sessions")
-            now_ts = time.time()
-            session_path_key = (
-                str(s.session_path.resolve())
-                if isinstance(s.session_path, Path)
-                else None
-            )
-            cache = getattr(self, "_pi_commands_cache", None)
-            cached = cache.get(runtime_id) if isinstance(cache, dict) else None
-            if isinstance(cached, dict):
-                cached_ts = cached.get("ts")
-                if (
-                    isinstance(cached_ts, (int, float))
-                    and (now_ts - float(cached_ts)) < PI_COMMANDS_CACHE_TTL_SECONDS
-                ):
-                    if (
-                        cached.get("thread_id") == s.thread_id
-                        and cached.get("session_path") == session_path_key
-                    ):
-                        commands = cached.get("commands")
-                        if isinstance(commands, list):
-                            return {"commands": list(commands)}
-            sock = s.sock_path
-            thread_id = s.thread_id
-        try:
-            resp = self._sock_call(sock, {"cmd": "commands"}, timeout_s=2.0)
-        except Exception:
-            if not _pid_alive(s.broker_pid) and not _pid_alive(s.codex_pid):
-                with self._lock:
-                    self._sessions.pop(runtime_id, None)
-                self._clear_deleted_session_state(runtime_id)
-                _unlink_quiet(sock)
-                _unlink_quiet(sock.with_suffix(".json"))
-                raise KeyError("unknown session")
-            raise ValueError("broker unavailable")
-
-        if resp.get("error") == "unknown cmd":
-            # Older live brokers may not expose the read-only commands RPC yet.
-            # Treat this as "feature unavailable" instead of surfacing a hard UI error.
-            return {"commands": []}
-
-        error = resp.get("error")
-        if isinstance(error, str) and error:
-            raise ValueError(error)
-        payload = _sanitize_pi_commands_payload(resp)
-        with self._lock:
-            cache = getattr(self, "_pi_commands_cache", None)
-            if isinstance(cache, dict):
-                cache[runtime_id] = {
-                    "ts": time.time(),
-                    "thread_id": thread_id,
-                    "session_path": session_path_key,
-                    "commands": list(payload.get("commands", [])),
-                }
-        return payload
+        return _pi_ui_bridge.get_session_commands(self, session_id)
 
     def submit_ui_response(
         self, session_id: str, payload: dict[str, Any]
     ) -> dict[str, Any]:
-        runtime_id = self._runtime_session_id_for_identifier(session_id)
-        if runtime_id is None:
-            raise KeyError("unknown session")
-        self.refresh_session_meta(runtime_id, strict=False)
-        with self._lock:
-            s = self._sessions.get(runtime_id)
-            if not s:
-                raise KeyError("unknown session")
-            if s.backend != "pi":
-                raise ValueError("ui interactions are only supported for pi sessions")
-            sock = s.sock_path
-        forward: dict[str, Any] = {"cmd": "ui_response"}
-        for key in ("id", "value", "confirmed", "cancelled"):
-            if key in payload:
-                forward[key] = payload[key]
-        try:
-            resp = self._sock_call(sock, forward, timeout_s=3.0)
-        except Exception:
-            if not _pid_alive(s.broker_pid) and not _pid_alive(s.codex_pid):
-                with self._lock:
-                    self._sessions.pop(runtime_id, None)
-                self._clear_deleted_session_state(runtime_id)
-                _unlink_quiet(sock)
-                _unlink_quiet(sock.with_suffix(".json"))
-                raise KeyError("unknown session")
-            raise ValueError("broker unavailable")
-        if resp.get("error") == "unknown cmd":
-            if _session_supports_live_pi_ui(s):
-                raise ValueError(
-                    "live ui responses are unavailable for this pi session"
-                )
-            if payload.get("cancelled") is True:
-                self.inject_keys(runtime_id, "\\x1b")
-                return {"ok": True, "legacy_fallback": True}
-            text = _legacy_pi_ui_response_text(payload)
-            if text is None:
-                raise ValueError("ui response value required")
-            self.send(runtime_id, text)
-            return {"ok": True, "legacy_fallback": True}
-        error = resp.get("error")
-        if isinstance(error, str) and error:
-            raise ValueError(error)
-        return resp
+        return _pi_ui_bridge.submit_ui_response(self, session_id, payload)
 
     def get_tail(self, session_id: str) -> str:
         runtime_id = self._runtime_session_id_for_identifier(session_id)
@@ -9311,1264 +8662,15 @@ class Handler(http.server.BaseHTTPRequestHandler):
                     self.send_error(404)
                     return
                 path = stripped
-            if path == "/favicon.ico":
-                resolved = _resolve_public_web_asset("favicon.ico")
-                if resolved is not None:
-                    self._send_path(resolved)
-                    return
-                self._send_static("favicon.png")
-                return
-            if path == "/manifest.webmanifest":
-                resolved = _resolve_public_web_asset("manifest.webmanifest")
-                if resolved is None:
-                    self.send_error(404)
-                    return
-                self._send_path(resolved)
-                return
-            if path == "/service-worker.js":
-                resolved = _resolve_public_web_asset("service-worker.js")
-                if resolved is None:
-                    self.send_error(404)
-                    return
-                self._send_path(resolved)
-                return
-            if path == "/app.js":
-                self._send_static("app.js")
-                return
-            if path == "/app.css":
-                self._send_static("app.css")
-                return
-            if path == "/favicon.png":
-                resolved = _resolve_public_web_asset("favicon.png")
-                if resolved is not None:
-                    self._send_path(resolved)
-                    return
-                self._send_static("favicon.png")
-                return
-            if path == "/":
-                body, ctype = _read_web_index()
-                self._send_bytes(body.encode("utf-8"), ctype)
-                return
-            if path.startswith("/assets/") and not USE_LEGACY_WEB:
-                served_dist_dir = _served_web_dist_dir()
-                if served_dist_dir is not None:
-                    candidate = (served_dist_dir / path.lstrip("/")).resolve()
-                    if (
-                        _is_path_within(served_dist_dir.resolve(), candidate)
-                        and candidate.is_file()
-                    ):
-                        self._send_path(candidate)
-                        return
-                self.send_error(404)
-                return
-            if (
-                not USE_LEGACY_WEB
-                and path.startswith("/")
-                and "/" not in path[1:]
-                and not path.startswith("/api/")
+            for route_module in (
+                _http_assets_routes,
+                _http_auth_routes,
+                _http_notification_routes,
+                _http_session_read_routes,
+                _http_file_routes,
             ):
-                resolved = _resolve_public_web_asset(path)
-                if resolved is not None:
-                    self._send_path(resolved)
+                if route_module.handle_get(self, path, u):
                     return
-            if path.startswith("/static/"):
-                self._send_static(path[len("/static/") :])
-                return
-
-            if path == "/api/me":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _json_response(self, 200, {"ok": True})
-                return
-
-            if path == "/api/settings/voice":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _json_response(
-                    self, 200, {"ok": True, **MANAGER._voice_push.settings_snapshot()}
-                )
-                return
-
-            if path == "/api/notifications/subscription":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {"ok": True, **MANAGER._voice_push.subscriptions_snapshot()},
-                )
-                return
-
-            if path == "/api/notifications/message":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                message_id = (qs.get("message_id") or [""])[0].strip()
-                if not message_id:
-                    _json_response(self, 400, {"error": "message_id required"})
-                    return
-                state = MANAGER._voice_push.notification_state_for_message(message_id)
-                if state is None:
-                    _json_response(self, 404, {"error": "unknown message"})
-                    return
-                _json_response(self, 200, {"ok": True, **state})
-                return
-
-            if path == "/api/notifications/feed":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                since_raw = (qs.get("since") or ["0"])[0].strip()
-                try:
-                    since_ts = float(since_raw or "0")
-                except ValueError:
-                    _json_response(self, 400, {"error": "invalid since"})
-                    return
-                items = MANAGER._voice_push.notification_feed_since(since_ts)
-                _json_response(self, 200, {"ok": True, "items": items})
-                return
-
-            if path == "/api/audio/live.m3u8":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = MANAGER._voice_push.playlist_bytes()
-                self.send_response(200)
-                self.send_header("Content-Type", "application/vnd.apple.mpegurl")
-                self.send_header("Content-Length", str(len(body)))
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(body)
-                return
-
-            if path.startswith("/api/audio/segments/"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                segment_name = path.split("/api/audio/segments/", 1)[1]
-                try:
-                    segment_path = MANAGER._voice_push.segment_path(segment_name)
-                except FileNotFoundError:
-                    self.send_error(404)
-                    return
-                raw = segment_path.read_bytes()
-                self.send_response(200)
-                self.send_header("Content-Type", "video/mp2t")
-                self.send_header("Content-Length", str(len(raw)))
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(raw)
-                return
-
-            if path == "/api/sessions/bootstrap":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "recent_cwds": MANAGER.recent_cwds(),
-                        "cwd_groups": MANAGER.cwd_groups_get(),
-                        "new_session_defaults": _read_new_session_defaults(),
-                        "tmux_available": _tmux_available(),
-                    },
-                )
-                return
-
-            if path == "/api/sessions":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                t0 = time.perf_counter()
-                qs = urllib.parse.parse_qs(u.query)
-                group_key_q = qs.get("group_key")
-                group_key = group_key_q[0] if group_key_q else None
-                offset = max(0, int(qs.get("offset", ["0"])[0] or "0"))
-                limit_default = SESSION_LIST_PAGE_SIZE
-                if group_key is not None:
-                    limit_default = SESSION_LIST_GROUP_PAGE_SIZE
-                limit = max(
-                    1,
-                    min(
-                        200,
-                        int(
-                            qs.get("limit", [str(limit_default)])[0]
-                            or str(limit_default)
-                        ),
-                    ),
-                )
-                group_offset = max(0, int(qs.get("group_offset", ["0"])[0] or "0"))
-                group_limit = max(
-                    1,
-                    min(
-                        20,
-                        int(
-                            qs.get(
-                                "group_limit", [str(SESSION_LIST_RECENT_GROUP_LIMIT)]
-                            )[0]
-                            or str(SESSION_LIST_RECENT_GROUP_LIMIT)
-                        ),
-                    ),
-                )
-                payload = _session_list_payload(
-                    MANAGER.list_sessions(),
-                    group_key=group_key,
-                    offset=offset,
-                    limit=limit,
-                    group_offset=group_offset,
-                    group_limit=group_limit,
-                )
-                dt_ms = (time.perf_counter() - t0) * 1000.0
-                _record_metric("api_sessions_ms", dt_ms)
-                _json_response(self, 200, payload)
-                return
-
-            if path == "/api/session_resume_candidates":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                cwd_raw = qs.get("cwd", [""])[0]
-                backend_raw = qs.get("backend", ["codex"])[0]
-                offset_raw = qs.get("offset", ["0"])[0]
-                limit_raw = qs.get("limit", ["20"])[0]
-                try:
-                    agent_backend = normalize_agent_backend(
-                        qs.get("agent_backend", [""])[0], default=DEFAULT_AGENT_BACKEND
-                    )
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                try:
-                    cwd_path = _resolve_dir_target(str(cwd_raw), field_name="cwd")
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e), "field": "cwd"})
-                    return
-                try:
-                    backend = _normalize_requested_backend(backend_raw)
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e), "field": "backend"})
-                    return
-                try:
-                    offset = max(0, int(offset_raw))
-                except ValueError:
-                    _json_response(self, 400, {"error": "offset must be an integer", "field": "offset"})
-                    return
-                try:
-                    limit = max(1, min(100, int(limit_raw)))
-                except ValueError:
-                    _json_response(self, 400, {"error": "limit must be an integer", "field": "limit"})
-                    return
-                info = _describe_session_cwd(cwd_path)
-                all_rows = (
-                    _list_resume_candidates_for_cwd(info["cwd"], backend=backend, limit=100000)
-                    if info["exists"]
-                    else []
-                )
-                rows = all_rows[offset : offset + limit]
-                remaining = max(0, len(all_rows) - (offset + len(rows)))
-                for row in rows:
-                    sid = row.get("session_id")
-                    alias = (
-                        MANAGER.alias_get(sid) if isinstance(sid, str) and sid else ""
-                    )
-                    preview = ""
-                    log_path_raw = row.get("log_path")
-                    session_path_raw = row.get("session_path")
-                    if isinstance(log_path_raw, str) and log_path_raw:
-                        preview = _first_user_message_preview_from_log(
-                            Path(log_path_raw)
-                        )
-                    elif isinstance(session_path_raw, str) and session_path_raw:
-                        preview = _first_user_message_preview_from_pi_session(
-                            Path(session_path_raw)
-                        )
-                    row["alias"] = alias
-                    row["first_user_message"] = preview
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        **info,
-                        "sessions": rows,
-                        "offset": offset,
-                        "limit": limit,
-                        "remaining": remaining,
-                    },
-                )
-                return
-
-            if path == "/api/metrics":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _json_response(self, 200, {"metrics": _metrics_snapshot()})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/live"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                offset_q = qs.get("offset")
-                live_offset_q = qs.get("live_offset")
-                bridge_offset_q = qs.get("bridge_offset")
-                requests_version_q = qs.get("requests_version")
-                if offset_q is None:
-                    offset = 0
-                else:
-                    if not offset_q:
-                        raise ValueError("invalid offset")
-                    offset = int(offset_q[0])
-                if live_offset_q is None:
-                    live_offset = 0
-                else:
-                    if not live_offset_q:
-                        raise ValueError("invalid live_offset")
-                    live_offset = int(live_offset_q[0])
-                if bridge_offset_q is None:
-                    bridge_offset = 0
-                else:
-                    if not bridge_offset_q:
-                        raise ValueError("invalid bridge_offset")
-                    bridge_offset = int(bridge_offset_q[0])
-                requests_version = None
-                if requests_version_q:
-                    requests_version = str(requests_version_q[0] or "").strip() or None
-                try:
-                    payload = _session_live_payload(
-                        MANAGER,
-                        session_id,
-                        offset=offset,
-                        live_offset=live_offset,
-                        bridge_offset=bridge_offset,
-                        requests_version=requests_version,
-                    )
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, payload)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/workspace"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                try:
-                    payload = _session_workspace_payload(MANAGER, session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, payload)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/details"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                try:
-                    payload = _session_details_payload(MANAGER, session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                _json_response(self, 200, payload)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/diagnostics"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                state = MANAGER.get_state(session_id)
-                if not isinstance(state, dict):
-                    raise ValueError("invalid broker state response")
-                if "busy" not in state:
-                    raise ValueError("missing busy from broker state response")
-                if "queue_len" not in state:
-                    raise ValueError("missing queue_len from broker state response")
-                token_val: dict[str, Any] | None = None
-                st_token = state.get("token")
-                if isinstance(st_token, dict) or st_token is None:
-                    token_val = (
-                        st_token
-                        if isinstance(st_token, dict)
-                        else (s.token if isinstance(s.token, dict) else None)
-                    )
-                model_provider = s.model_provider
-                preferred_auth_method = s.preferred_auth_method
-                model = s.model
-                reasoning_effort = s.reasoning_effort
-                service_tier = s.service_tier
-                if (
-                    (
-                        model_provider is None
-                        or model is None
-                        or reasoning_effort is None
-                    )
-                    and s.log_path is not None
-                    and s.log_path.exists()
-                ):
-                    log_provider, log_model, log_effort = _read_run_settings_from_log(
-                        s.log_path, agent_backend=s.agent_backend
-                    )
-                    if model_provider is None:
-                        model_provider = log_provider
-                    if model is None:
-                        model = log_model
-                    if reasoning_effort is None:
-                        reasoning_effort = log_effort
-                sidebar_meta = MANAGER.sidebar_meta_get(session_id)
-                cwd_path = _safe_expanduser(Path(s.cwd))
-                if not cwd_path.is_absolute():
-                    cwd_path = cwd_path.resolve()
-                git_branch = _current_git_branch(cwd_path)
-                updated_ts = _display_updated_ts(s)
-                elapsed_s = max(0.0, time.time() - updated_ts)
-                time_priority = _priority_from_elapsed_seconds(elapsed_s)
-                base_priority = _clip01(
-                    time_priority + float(sidebar_meta["priority_offset"])
-                )
-                blocked = sidebar_meta["dependency_session_id"] is not None
-                snoozed = (
-                    sidebar_meta["snooze_until"] is not None
-                    and float(sidebar_meta["snooze_until"]) > time.time()
-                )
-                final_priority = 0.0 if (snoozed or blocked) else base_priority
-                broker_busy = bool(state.get("busy"))
-                busy = (
-                    _display_pi_busy(s, broker_busy=broker_busy)
-                    if s.backend == "pi"
-                    else broker_busy
-                )
-                if s.backend != "pi" and s.log_path is not None and s.log_path.exists():
-                    idle_val = MANAGER.idle_from_log(session_id)
-                    busy = broker_busy or (not bool(idle_val))
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "session_id": _durable_session_id_for_live_session(s),
-                        "runtime_id": s.session_id,
-                        "thread_id": s.thread_id,
-                        "agent_backend": s.agent_backend,
-                        "backend": s.backend,
-                        "agent_backend": s.agent_backend,
-                        "owned": bool(s.owned),
-                        "transport": s.transport,
-                        "cwd": s.cwd,
-                        "start_ts": float(s.start_ts),
-                        "updated_ts": updated_ts,
-                        "log_path": (
-                            str(s.log_path) if s.log_path is not None else None
-                        ),
-                        "session_file_path": _display_source_path(s),
-                        "broker_pid": int(s.broker_pid),
-                        "codex_pid": int(s.codex_pid),
-                        "busy": bool(busy),
-                        "broker_busy": broker_busy,
-                        "queue_len": MANAGER._queue_len(session_id),
-                        "token": token_val,
-                        "model_provider": model_provider,
-                        "preferred_auth_method": preferred_auth_method,
-                        "provider_choice": _provider_choice_for_backend(
-                            backend=s.backend,
-                            model_provider=model_provider,
-                            preferred_auth_method=preferred_auth_method,
-                        ),
-                        "model": model,
-                        "reasoning_effort": reasoning_effort,
-                        "service_tier": service_tier,
-                        "tmux_session": s.tmux_session,
-                        "tmux_window": s.tmux_window,
-                        "git_branch": git_branch,
-                        "time_priority": time_priority,
-                        "base_priority": base_priority,
-                        "final_priority": final_priority,
-                        "priority_offset": sidebar_meta["priority_offset"],
-                        "snooze_until": sidebar_meta["snooze_until"],
-                        "dependency_session_id": sidebar_meta["dependency_session_id"],
-                        "todo_snapshot": _todo_snapshot_payload_for_session(s),
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/queue"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                try:
-                    q = MANAGER.queue_list(session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, "queue": q})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/ui_state"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                try:
-                    payload = MANAGER.get_ui_state(session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, payload)
-                return
-
-            session_id = _match_session_route(path, "commands")
-            if session_id is not None:
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                try:
-                    payload = MANAGER.get_session_commands(session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, payload)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/file/read"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                rel = path_q[0]
-                base = _safe_expanduser(Path(s.cwd))
-                if not base.is_absolute():
-                    base = base.resolve()
-                p = _resolve_session_path(base, rel)
-                if not p.exists():
-                    _json_response(self, 404, {"error": "file not found"})
-                    return
-                if not p.is_file():
-                    _json_response(self, 400, {"error": "path is not a file"})
-                    return
-                try:
-                    view = _read_client_file_view(p)
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                try:
-                    MANAGER.files_add(session_id, str(p))
-                except KeyError:
-                    pass
-                if view.kind == "image":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "image",
-                            "content_type": view.content_type,
-                            "path": str(p),
-                            "rel": str(rel),
-                            "size": int(view.size),
-                            "image_url": f"/api/sessions/{session_id}/file/blob?path={urllib.parse.quote(rel)}",
-                        },
-                    )
-                    return
-                if view.kind == "pdf":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "pdf",
-                            "content_type": view.content_type,
-                            "path": str(p),
-                            "rel": str(rel),
-                            "size": int(view.size),
-                            "pdf_url": f"/api/sessions/{session_id}/file/blob?path={urllib.parse.quote(rel)}",
-                        },
-                    )
-                    return
-                if view.kind == "download_only":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "download_only",
-                            "path": str(p),
-                            "rel": str(rel),
-                            "size": int(view.size),
-                            "reason": view.blocked_reason,
-                            "viewer_max_bytes": view.viewer_max_bytes,
-                        },
-                    )
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "kind": view.kind,
-                        "path": str(p),
-                        "rel": str(rel),
-                        "size": int(view.size),
-                        "text": view.text,
-                        "editable": bool(view.editable),
-                        "version": view.version,
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/file/search"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                query_raw = qs.get("q")
-                if not query_raw or not query_raw[0].strip():
-                    _json_response(self, 400, {"error": "q required"})
-                    return
-                limit_raw = qs.get("limit", [str(FILE_SEARCH_LIMIT)])[0]
-                try:
-                    limit = int(str(limit_raw).strip() or str(FILE_SEARCH_LIMIT))
-                except ValueError:
-                    _json_response(self, 400, {"error": "limit must be an integer"})
-                    return
-                if limit < 1:
-                    _json_response(self, 400, {"error": "limit must be >= 1"})
-                    return
-                base = _safe_expanduser(Path(s.cwd))
-                if not base.is_absolute():
-                    base = base.resolve()
-                try:
-                    result = _search_session_relative_files(
-                        base, query=query_raw[0], limit=limit
-                    )
-                except FileNotFoundError as e:
-                    _json_response(self, 404, {"error": str(e)})
-                    return
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except (RuntimeError, ValueError) as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "cwd": str(base),
-                        "query": result["query"],
-                        "mode": result["mode"],
-                        "matches": result["matches"],
-                        "scanned": result["scanned"],
-                        "truncated": result["truncated"],
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/file/list"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                raw_rel = qs.get("path", [""])[0]
-                base = _safe_expanduser(Path(s.cwd))
-                if not base.is_absolute():
-                    base = base.resolve()
-                try:
-                    entries = _list_session_directory_entries(base, raw_rel)
-                except FileNotFoundError as e:
-                    _json_response(self, 404, {"error": str(e)})
-                    return
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "cwd": str(base),
-                        "path": str(raw_rel or ""),
-                        "entries": entries,
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/file/blob"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                rel = path_q[0]
-                base = _safe_expanduser(Path(s.cwd))
-                if not base.is_absolute():
-                    base = base.resolve()
-                p = _resolve_session_path(base, rel)
-                if not p.exists():
-                    _json_response(self, 404, {"error": "file not found"})
-                    return
-                if not p.is_file():
-                    _json_response(self, 400, {"error": "path is not a file"})
-                    return
-                raw = p.read_bytes()
-                _kind, ctype = _file_kind(p, raw)
-                if _kind not in {"image", "pdf"} or ctype is None:
-                    _json_response(
-                        self, 400, {"error": "file is not previewable inline"}
-                    )
-                    return
-                self.send_response(200)
-                self.send_header("Content-Type", ctype)
-                self.send_header("Content-Length", str(len(raw)))
-                self.send_header(
-                    "Content-Disposition",
-                    f"inline; filename*=UTF-8''{urllib.parse.quote(p.name, safe='')}",
-                )
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(raw)
-                return
-
-            if path == "/api/files/blob":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                path_obj = Path(path_q[0]).expanduser().resolve()
-                if not path_obj.exists():
-                    _json_response(self, 404, {"error": "file not found"})
-                    return
-                if not path_obj.is_file():
-                    _json_response(self, 400, {"error": "path is not a file"})
-                    return
-                raw = path_obj.read_bytes()
-                _kind, ctype = _file_kind(path_obj, raw)
-                if _kind not in {"image", "pdf"} or ctype is None:
-                    _json_response(
-                        self, 400, {"error": "file is not previewable inline"}
-                    )
-                    return
-                self.send_response(200)
-                self.send_header("Content-Type", ctype)
-                self.send_header("Content-Length", str(len(raw)))
-                self.send_header(
-                    "Content-Disposition",
-                    f"inline; filename*=UTF-8''{urllib.parse.quote(path_obj.name, safe='')}",
-                )
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(raw)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/file/download"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                rel = path_q[0]
-                base = _safe_expanduser(Path(s.cwd))
-                if not base.is_absolute():
-                    base = base.resolve()
-                p = _resolve_session_path(base, rel)
-                try:
-                    raw, size = _read_downloadable_file(p)
-                except FileNotFoundError as e:
-                    _json_response(self, 404, {"error": str(e)})
-                    return
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                self.send_response(200)
-                self.send_header("Content-Type", "application/octet-stream")
-                self.send_header("Content-Length", str(size))
-                self.send_header("Content-Disposition", _download_disposition(p))
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(raw)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith(
-                "/git/changed_files"
-            ):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                cwd = _safe_expanduser(Path(s.cwd))
-                if not cwd.is_absolute():
-                    cwd = cwd.resolve()
-                try:
-                    _require_git_repo(cwd)
-                except RuntimeError as e:
-                    _json_response(self, 409, {"error": str(e)})
-                    return
-                unstaged = _run_git(
-                    cwd,
-                    ["diff", "--name-only"],
-                    timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                    max_bytes=64 * 1024,
-                ).splitlines()
-                staged = _run_git(
-                    cwd,
-                    ["diff", "--name-only", "--cached"],
-                    timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                    max_bytes=64 * 1024,
-                ).splitlines()
-                unstaged_numstat = _run_git(
-                    cwd,
-                    ["diff", "--numstat"],
-                    timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                    max_bytes=128 * 1024,
-                )
-                staged_numstat = _run_git(
-                    cwd,
-                    ["diff", "--numstat", "--cached"],
-                    timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                    max_bytes=128 * 1024,
-                )
-
-                def _norm_list(xs: list[str]) -> list[str]:
-                    out: list[str] = []
-                    for x in xs:
-                        t = x.strip()
-                        if not t:
-                            continue
-                        out.append(t)
-                        if len(out) >= GIT_CHANGED_FILES_MAX:
-                            break
-                    return out
-
-                unstaged2 = _norm_list(unstaged)
-                staged2 = _norm_list(staged)
-                seen: set[str] = set()
-                merged: list[str] = []
-                for x in [*unstaged2, *staged2]:
-                    if x in seen:
-                        continue
-                    seen.add(x)
-                    merged.append(x)
-                stats = _parse_git_numstat(unstaged_numstat)
-                for path_key, vals in _parse_git_numstat(staged_numstat).items():
-                    prev = stats.get(path_key)
-                    if prev is None:
-                        stats[path_key] = vals
-                        continue
-                    add_prev = prev.get("additions")
-                    del_prev = prev.get("deletions")
-                    add_new = vals.get("additions")
-                    del_new = vals.get("deletions")
-                    prev["additions"] = (
-                        None
-                        if add_prev is None or add_new is None
-                        else int(add_prev) + int(add_new)
-                    )
-                    prev["deletions"] = (
-                        None
-                        if del_prev is None or del_new is None
-                        else int(del_prev) + int(del_new)
-                    )
-                entries: list[dict[str, Any]] = []
-                for path_key in merged:
-                    vals = stats.get(path_key, {})
-                    entries.append(
-                        {
-                            "path": path_key,
-                            "additions": vals.get("additions"),
-                            "deletions": vals.get("deletions"),
-                            "changed": True,
-                        }
-                    )
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "cwd": str(cwd),
-                        "files": merged,
-                        "entries": entries,
-                        "unstaged": unstaged2,
-                        "staged": staged2,
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/git/diff"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                rel = path_q[0]
-                staged_q = qs.get("staged")
-                staged = bool(staged_q and staged_q[0] == "1")
-                cwd = _safe_expanduser(Path(s.cwd))
-                if not cwd.is_absolute():
-                    cwd = cwd.resolve()
-                try:
-                    _require_git_repo(cwd)
-                except RuntimeError as e:
-                    _json_response(self, 409, {"error": str(e)})
-                    return
-                try:
-                    _target, _repo_root, rel = _resolve_git_path(cwd, rel)
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                args = ["diff", "-U3"]
-                if staged:
-                    args.append("--cached")
-                args.extend(["--", rel])
-                diff = _run_git(
-                    cwd,
-                    args,
-                    timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                    max_bytes=GIT_DIFF_MAX_BYTES,
-                )
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "cwd": str(cwd),
-                        "path": rel,
-                        "staged": staged,
-                        "diff": diff,
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith(
-                "/git/file_versions"
-            ):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                if not session_id:
-                    self.send_error(404)
-                    return
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                rel = path_q[0]
-                cwd = _safe_expanduser(Path(s.cwd))
-                if not cwd.is_absolute():
-                    cwd = cwd.resolve()
-                try:
-                    _require_git_repo(cwd)
-                except RuntimeError as e:
-                    _json_response(self, 409, {"error": str(e)})
-                    return
-                try:
-                    p, _repo_root, rel = _resolve_git_path(cwd, rel)
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                current_text = ""
-                current_size = 0
-                current_exists = bool(p.exists() and p.is_file())
-                if current_exists:
-                    current_text, current_size = _read_text_file_strict(
-                        p, max_bytes=FILE_READ_MAX_BYTES
-                    )
-                try:
-                    MANAGER.files_add(session_id, str(p))
-                except KeyError:
-                    pass
-                base_exists = False
-                base_text = ""
-                try:
-                    base_text = _run_git(
-                        cwd,
-                        ["show", f"HEAD:{rel}"],
-                        timeout_s=GIT_DIFF_TIMEOUT_SECONDS,
-                        max_bytes=FILE_READ_MAX_BYTES,
-                    )
-                    base_exists = True
-                except RuntimeError:
-                    base_exists = False
-                    base_text = ""
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "cwd": str(cwd),
-                        "path": rel,
-                        "abs_path": str(p),
-                        "current_exists": current_exists,
-                        "current_size": int(current_size),
-                        "current_text": current_text,
-                        "base_exists": base_exists,
-                        "base_text": base_text,
-                    },
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/messages"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                t0_total = time.perf_counter()
-                parts = path.split("/")
-                if len(parts) < 4:
-                    self.send_error(404)
-                    return
-                session_id = parts[3]
-                t0_meta = time.perf_counter()
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                dt_meta_ms = (time.perf_counter() - t0_meta) * 1000.0
-                s = MANAGER.get_session(session_id)
-                historical_row = _historical_session_row(session_id)
-                if (not s) and historical_row is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                offset_q = qs.get("offset")
-                if offset_q is None:
-                    offset = 0
-                else:
-                    if not offset_q:
-                        raise ValueError("invalid offset")
-                    offset = int(offset_q[0])
-                if offset < 0:
-                    offset = 0
-                init_q = qs.get("init")
-                init = bool(init_q and init_q[0] == "1")
-                before_q = qs.get("before")
-                if before_q is None:
-                    before = 0
-                else:
-                    if not before_q:
-                        raise ValueError("invalid before")
-                    before = int(before_q[0])
-                before = max(0, before)
-                limit_q = qs.get("limit")
-                if limit_q is None:
-                    limit = 80
-                else:
-                    if not limit_q:
-                        raise ValueError("invalid limit")
-                    limit = int(limit_q[0])
-                limit = max(20, min(200, limit))
-                payload = MANAGER.get_messages_page(
-                    session_id,
-                    offset=offset,
-                    init=init,
-                    limit=limit,
-                    before=before,
-                )
-                if isinstance(payload.get("diag"), dict) and s is not None and s.backend != "pi":
-                    payload["diag"]["meta_refresh_ms"] = round(dt_meta_ms, 3)
-                _json_response(self, 200, payload)
-                dt_total_ms = (time.perf_counter() - t0_total) * 1000.0
-                _record_metric(
-                    "api_messages_init_ms" if init else "api_messages_poll_ms",
-                    dt_total_ms,
-                )
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/tail"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                try:
-                    tail = MANAGER.get_tail(session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                _json_response(self, 200, {"tail": tail})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/harness"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                if len(parts) < 4:
-                    self.send_error(404)
-                    return
-                session_id = parts[3]
-                try:
-                    cfg = MANAGER.harness_get(session_id)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                _json_response(self, 200, {"ok": True, **cfg})
-                return
-
             self.send_error(404)
         except Exception as e:
             traceback.print_exc()
@@ -10594,996 +8696,14 @@ class Handler(http.server.BaseHTTPRequestHandler):
                     self.send_error(404)
                     return
                 path = stripped
-
-            if path == "/api/cwd_groups/edit":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                try:
-                    body = _read_body(self)
-                    body_text = body.decode("utf-8")
-                    if not body_text.strip():
-                        raise ValueError("empty request body")
-                    obj = json.loads(body_text)
-                    if not isinstance(obj, dict):
-                        raise ValueError("invalid json body (expected object)")
-                    cwd, entry = MANAGER.cwd_group_set(
-                        cwd=obj.get("cwd"),
-                        label=obj.get("label"),
-                        collapsed=obj.get("collapsed"),
-                    )
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, "cwd": cwd, **entry})
-                return
-
-            if path == "/api/login":
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                pw = obj.get("password")
-                if not isinstance(pw, str) or not _is_same_password(pw):
-                    _json_response(self, 403, {"error": "bad password"})
-                    return
-                self.send_response(200)
-                _set_auth_cookie(self)
-                self.send_header("Content-Type", "application/json; charset=utf-8")
-                self.end_headers()
-                self.wfile.write(b'{"ok":true}')
-                return
-
-            if path == "/api/logout":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                self.send_response(200)
-                self.send_header(
-                    "Set-Cookie",
-                    f"{COOKIE_NAME}=deleted; Path={COOKIE_PATH}; Max-Age=0; HttpOnly; SameSite=Strict",
-                )
-                self.send_header("Content-Type", "application/json; charset=utf-8")
-                self.end_headers()
-                self.wfile.write(b'{"ok":true}')
-                return
-
-            if path == "/api/settings/voice":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                try:
-                    payload = MANAGER._voice_push.set_settings(obj)
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/notifications/subscription":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                try:
-                    payload = MANAGER._voice_push.upsert_subscription(
-                        subscription=obj.get("subscription"),
-                        user_agent=str(obj.get("user_agent") or ""),
-                        device_label=str(obj.get("device_label") or ""),
-                        device_class=str(obj.get("device_class") or ""),
-                    )
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/notifications/subscription/toggle":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                endpoint = obj.get("endpoint")
-                enabled = obj.get("enabled")
-                if not isinstance(endpoint, str) or not endpoint.strip():
-                    _json_response(self, 400, {"error": "endpoint required"})
-                    return
-                if not isinstance(enabled, bool):
-                    _json_response(self, 400, {"error": "enabled must be a boolean"})
-                    return
-                try:
-                    payload = MANAGER._voice_push.toggle_subscription(
-                        endpoint=endpoint, enabled=enabled
-                    )
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown subscription"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/notifications/test_push":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                try:
-                    payload = MANAGER._voice_push.send_test_push_notification(
-                        session_display_name="Codoxear test"
-                    )
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/audio/listener":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                client_id = obj.get("client_id")
-                enabled = obj.get("enabled")
-                if not isinstance(client_id, str) or not client_id.strip():
-                    _json_response(self, 400, {"error": "client_id required"})
-                    return
-                if not isinstance(enabled, bool):
-                    _json_response(self, 400, {"error": "enabled must be a boolean"})
-                    return
-                payload = MANAGER._voice_push.listener_heartbeat(
-                    client_id=client_id, enabled=enabled
-                )
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/audio/test_announcement":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                try:
-                    payload = MANAGER._voice_push.enqueue_test_announcement(
-                        session_display_name="Codoxear test"
-                    )
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, **payload})
-                return
-
-            if path == "/api/sessions":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                try:
-                    payload = _parse_create_session_request(obj)
-                except ValueError as e:
-                    err = str(e)
-                    out: dict[str, Any] = {"error": err}
-                    if err == "cwd required":
-                        out["field"] = "cwd"
-                    _json_response(self, 400, out)
-                    return
-                try:
-                    res = MANAGER.spawn_web_session(
-                        cwd=payload["cwd"],
-                        args=payload["args"],
-                        resume_session_id=payload["resume_session_id"],
-                        worktree_branch=payload["worktree_branch"],
-                        model_provider=payload["model_provider"],
-                        preferred_auth_method=payload["preferred_auth_method"],
-                        model=payload["model"],
-                        reasoning_effort=payload["reasoning_effort"],
-                        service_tier=payload["service_tier"],
-                        create_in_tmux=payload["create_in_tmux"],
-                        backend=payload["backend"],
-                    )
-                    alias = MANAGER.set_created_session_name(
-                        session_id=res.get("session_id"),
-                        runtime_id=res.get("runtime_id"),
-                        backend=res.get("backend") or payload["backend"],
-                        name=payload["name"],
-                    )
-                except ValueError as e:
-                    payload = {"error": str(e)}
-                    if str(e).startswith("cwd "):
-                        payload["field"] = "cwd"
-                    _json_response(self, 400, payload)
-                    return
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                response_payload = {"ok": True, **res}
-                if alias:
-                    response_payload["alias"] = alias
-                _json_response(self, 200, response_payload)
-                return
-
-            if path == "/api/files/read":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                path_raw = obj.get("path")
-                if not isinstance(path_raw, str) or not path_raw.strip():
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                session_id_raw = obj.get("session_id")
-                session_id = (
-                    session_id_raw
-                    if isinstance(session_id_raw, str) and session_id_raw
-                    else ""
-                )
-                try:
-                    path_obj = _resolve_client_file_path(
-                        session_id=session_id, raw_path=path_raw
-                    )
-                    view = _read_client_file_view(path_obj)
-                except FileNotFoundError as e:
-                    _json_response(self, 404, {"error": str(e)})
-                    return
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                if session_id:
-                    try:
-                        MANAGER.files_add(session_id, str(path_obj))
-                    except KeyError:
-                        pass
-                if view.kind == "image":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "image",
-                            "content_type": view.content_type,
-                            "path": str(path_obj),
-                            "size": int(view.size),
-                            "image_url": f"/api/files/blob?path={urllib.parse.quote(str(path_obj))}",
-                        },
-                    )
-                    return
-                if view.kind == "pdf":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "pdf",
-                            "content_type": view.content_type,
-                            "path": str(path_obj),
-                            "size": int(view.size),
-                            "pdf_url": f"/api/files/blob?path={urllib.parse.quote(str(path_obj))}",
-                        },
-                    )
-                    return
-                if view.kind == "download_only":
-                    _json_response(
-                        self,
-                        200,
-                        {
-                            "ok": True,
-                            "kind": "download_only",
-                            "path": str(path_obj),
-                            "size": int(view.size),
-                            "reason": view.blocked_reason,
-                            "viewer_max_bytes": view.viewer_max_bytes,
-                        },
-                    )
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "kind": view.kind,
-                        "path": str(path_obj),
-                        "size": int(view.size),
-                        "text": view.text,
-                        "editable": bool(view.editable),
-                        "version": view.version,
-                    },
-                )
-                return
-
-            if path == "/api/files/inspect":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                path_raw = obj.get("path")
-                if not isinstance(path_raw, str) or not path_raw.strip():
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                session_id_raw = obj.get("session_id")
-                session_id = (
-                    session_id_raw
-                    if isinstance(session_id_raw, str) and session_id_raw
-                    else ""
-                )
-                try:
-                    path_obj = _resolve_client_file_path(
-                        session_id=session_id, raw_path=path_raw
-                    )
-                    view = _read_client_file_view(path_obj)
-                except FileNotFoundError as e:
-                    _json_response(self, 404, {"error": str(e)})
-                    return
-                except PermissionError as e:
-                    _json_response(self, 403, {"error": str(e)})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "path": str(path_obj),
-                        "kind": view.kind,
-                        "content_type": view.content_type,
-                        "size": int(view.size),
-                        "reason": view.blocked_reason,
-                        "viewer_max_bytes": view.viewer_max_bytes,
-                    },
-                )
-                return
-
-            if path == "/api/files/blob":
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                qs = urllib.parse.parse_qs(u.query)
-                path_q = qs.get("path")
-                if not path_q or not path_q[0]:
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                path_obj = _safe_expanduser(Path(path_q[0])).resolve()
-                if not path_obj.exists():
-                    _json_response(self, 404, {"error": "file not found"})
-                    return
-                if not path_obj.is_file():
-                    _json_response(self, 400, {"error": "path is not a file"})
-                    return
-                raw = path_obj.read_bytes()
-                _kind, ctype = _file_kind(path_obj, raw)
-                if _kind not in {"image", "pdf"} or ctype is None:
-                    _json_response(
-                        self, 400, {"error": "file is not previewable inline"}
-                    )
-                    return
-                self.send_response(200)
-                self.send_header("Content-Type", ctype)
-                self.send_header("Content-Length", str(len(raw)))
-                self.send_header(
-                    "Content-Disposition",
-                    f"inline; filename*=UTF-8''{urllib.parse.quote(path_obj.name, safe='')}",
-                )
-                self.send_header("Cache-Control", "no-store")
-                self.send_header("Pragma", "no-cache")
-                self.send_header("Expires", "0")
-                self.end_headers()
-                self.wfile.write(raw)
-                return
-
-            session_id = _match_session_route(path, "file", "write")
-            if session_id is not None:
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                path_raw = obj.get("path")
-                if not isinstance(path_raw, str) or not path_raw.strip():
-                    _json_response(self, 400, {"error": "path required"})
-                    return
-                text_raw = obj.get("text")
-                if not isinstance(text_raw, str):
-                    _json_response(self, 400, {"error": "text must be a string"})
-                    return
-                create_raw = obj.get("create")
-                create = create_raw if isinstance(create_raw, bool) else False
-                version_raw = obj.get("version")
-                if not create and (
-                    not isinstance(version_raw, str) or not version_raw.strip()
-                ):
-                    _json_response(self, 400, {"error": "version required"})
-                    return
-                MANAGER.refresh_session_meta(session_id)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                base = Path(s.cwd).expanduser()
-                if not base.is_absolute():
-                    base = base.resolve()
-                if create:
-                    try:
-                        p = _resolve_under(base, path_raw)
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                    try:
-                        size, next_version = _write_new_text_file_atomic(
-                            p, text=text_raw
-                        )
-                    except FileExistsError:
-                        payload: dict[str, Any] = {
-                            "error": "file already exists",
-                            "conflict": True,
-                            "path": str(p),
-                        }
-                        if p.is_file():
-                            try:
-                                _current_text, _current_size, current_version = (
-                                    _read_text_file_for_write(
-                                        p, max_bytes=FILE_READ_MAX_BYTES
-                                    )
-                                )
-                                payload["version"] = current_version
-                            except (FileNotFoundError, PermissionError, ValueError):
-                                pass
-                        _json_response(self, 409, payload)
-                        return
-                    except FileNotFoundError as e:
-                        _json_response(self, 404, {"error": str(e)})
-                        return
-                    except PermissionError as e:
-                        _json_response(self, 403, {"error": str(e)})
-                        return
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                else:
-                    p = _resolve_session_path(base, path_raw)
-                    try:
-                        _current_text, _current_size, current_version = (
-                            _read_text_file_for_write(p, max_bytes=FILE_READ_MAX_BYTES)
-                        )
-                    except FileNotFoundError as e:
-                        _json_response(self, 404, {"error": str(e)})
-                        return
-                    except PermissionError as e:
-                        _json_response(self, 403, {"error": str(e)})
-                        return
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                    if current_version != version_raw:
-                        _json_response(
-                            self,
-                            409,
-                            {
-                                "error": "file changed on disk",
-                                "conflict": True,
-                                "path": str(p),
-                                "version": current_version,
-                            },
-                        )
-                        return
-                    try:
-                        size, next_version = _write_text_file_atomic(p, text=text_raw)
-                    except FileNotFoundError as e:
-                        _json_response(self, 404, {"error": str(e)})
-                        return
-                    except PermissionError as e:
-                        _json_response(self, 403, {"error": str(e)})
-                        return
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                try:
-                    MANAGER.files_add(session_id, str(p))
-                except KeyError:
-                    pass
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "path": str(p),
-                        "rel": str(path_raw),
-                        "size": int(size),
-                        "version": next_version,
-                        "editable": True,
-                    },
-                )
-                return
-
-            session_id = _match_session_route(path, "delete")
-            if session_id is not None:
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                _read_body(self)
-                ok = MANAGER.delete_session(session_id)
-                if not ok:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                _json_response(self, 200, {"ok": True})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/edit"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                name = obj.get("name")
-                if not isinstance(name, str):
-                    _json_response(self, 400, {"error": "name required"})
-                    return
-                try:
-                    alias, sidebar_meta = MANAGER.edit_session(
-                        session_id,
-                        name=name,
-                        priority_offset=obj.get("priority_offset"),
-                        snooze_until=obj.get("snooze_until"),
-                        dependency_session_id=obj.get("dependency_session_id"),
-                    )
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, "alias": alias, **sidebar_meta})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/rename"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                session_id = _match_session_route(path, "rename")
-                if session_id is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                name = obj.get("name")
-                if not isinstance(name, str):
-                    _json_response(self, 400, {"error": "name required"})
-                    return
-                try:
-                    alias = MANAGER.alias_set(session_id, name)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                _json_response(self, 200, {"ok": True, "alias": alias})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/focus"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                session_id = _match_session_route(path, "focus")
-                if session_id is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                try:
-                    focused = MANAGER.focus_set(session_id, obj.get("focused"))
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, "focused": focused})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/send"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                session_id = _match_session_route(path, "send")
-                if session_id is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                text = obj.get("text")
-                if not isinstance(text, str) or not text.strip():
-                    _json_response(self, 400, {"error": "text required"})
-                    return
-                try:
-                    res = MANAGER.send(session_id, text)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, res)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/ui_response"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                session_id = _match_session_route(path, "ui_response")
-                if session_id is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                try:
-                    MANAGER.submit_ui_response(session_id, obj)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/enqueue"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                session_id = _match_session_route(path, "enqueue")
-                if session_id is None:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                text = obj.get("text")
-                if not isinstance(text, str) or not text.strip():
-                    _json_response(self, 400, {"error": "text required"})
-                    return
-                try:
-                    res = MANAGER.enqueue(session_id, text)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, res)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/queue/delete"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                idx = obj.get("index")
-                if not isinstance(idx, int):
-                    _json_response(self, 400, {"error": "index required"})
-                    return
-                try:
-                    res = MANAGER.queue_delete(session_id, idx)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, res)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/queue/update"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                idx = obj.get("index")
-                text = obj.get("text")
-                if not isinstance(idx, int):
-                    _json_response(self, 400, {"error": "index required"})
-                    return
-                if not isinstance(text, str) or not text.strip():
-                    _json_response(self, 400, {"error": "text required"})
-                    return
-                try:
-                    res = MANAGER.queue_update(session_id, idx, text)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, res)
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/harness"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                body = _read_body(self)
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                enabled_raw = obj.get("enabled", None)
-                request_raw = obj.get("request", None)
-                cooldown_minutes_raw = obj.get("cooldown_minutes", None)
-                remaining_injections_raw = obj.get("remaining_injections", None)
-                if "text" in obj:
-                    _json_response(
-                        self, 400, {"error": "unknown field: text (use request)"}
-                    )
-                    return
-                enabled: bool | None
-                if enabled_raw is None:
-                    enabled = None
-                else:
-                    enabled = bool(enabled_raw)
-
-                if request_raw is not None and (not isinstance(request_raw, str)):
-                    _json_response(self, 400, {"error": "request must be a string"})
-                    return
-                request: str | None
-                if request_raw is not None:
-                    request = request_raw
-                else:
-                    request = None
-                cooldown_minutes: int | None
-                if cooldown_minutes_raw is not None:
-                    try:
-                        cooldown_minutes = _clean_harness_cooldown_minutes(
-                            cooldown_minutes_raw
-                        )
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                else:
-                    cooldown_minutes = None
-                remaining_injections: int | None
-                if remaining_injections_raw is not None:
-                    try:
-                        remaining_injections = _clean_harness_remaining_injections(
-                            remaining_injections_raw, allow_zero=True
-                        )
-                    except ValueError as e:
-                        _json_response(self, 400, {"error": str(e)})
-                        return
-                else:
-                    remaining_injections = None
-
-                cfg = MANAGER.harness_set(
-                    session_id,
-                    enabled=enabled,
-                    request=request,
-                    cooldown_minutes=cooldown_minutes,
-                    remaining_injections=remaining_injections,
-                )
-                _json_response(self, 200, {"ok": True, **cfg})
-                return
-
-            if path.startswith("/api/sessions/") and path.endswith("/interrupt"):
-                if not _require_auth(self):
-                    self._unauthorized()
-                    return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                _read_body(self)
-                try:
-                    # Send a literal ESC byte. Older brokers may not recognize "ESC" but will
-                    # decode "\\x1b" via unicode_escape into a single 0x1b byte.
-                    resp = MANAGER.inject_keys(session_id, "\\x1b")
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(self, 200, {"ok": True, "broker": resp})
-                return
-
-            if path.startswith("/api/sessions/") and (
-                path.endswith("/inject_file") or path.endswith("/inject_image")
+            for route_module in (
+                _http_auth_routes,
+                _http_notification_routes,
+                _http_session_write_routes,
+                _http_file_routes,
             ):
-                if not _require_auth(self):
-                    self._unauthorized()
+                if route_module.handle_post(self, path, u):
                     return
-                parts = path.split("/")
-                session_id = parts[3] if len(parts) >= 4 else ""
-                MANAGER.refresh_session_meta(session_id, strict=False)
-                s = MANAGER.get_session(session_id)
-                if not s:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                if s.backend == "pi":
-                    _json_response(
-                        self,
-                        409,
-                        {
-                            "error": "attachment injection is not supported for Pi sessions",
-                            "backend": "pi",
-                            "operation": "attachment_injection",
-                        },
-                    )
-                    return
-                try:
-                    body = _read_body(self, limit=ATTACH_UPLOAD_BODY_MAX_BYTES)
-                except ValueError:
-                    _json_response(
-                        self,
-                        413,
-                        {
-                            "error": f"file too large (max {ATTACH_UPLOAD_MAX_BYTES} bytes)"
-                        },
-                    )
-                    return
-                body_text = body.decode("utf-8")
-                if not body_text.strip():
-                    raise ValueError("empty request body")
-                obj = json.loads(body_text)
-                if not isinstance(obj, dict):
-                    raise ValueError("invalid json body (expected object)")
-                data_b64 = obj.get("data_b64")
-                filename = obj.get("filename")
-                attachment_index = obj.get("attachment_index")
-                if not isinstance(filename, str) or (not filename.strip()):
-                    raise ValueError("filename required")
-                if isinstance(attachment_index, bool) or not isinstance(
-                    attachment_index, int
-                ):
-                    _json_response(
-                        self, 400, {"error": "attachment_index must be an integer"}
-                    )
-                    return
-                if not isinstance(data_b64, str) or not data_b64:
-                    _json_response(self, 400, {"error": "data_b64 required"})
-                    return
-                try:
-                    raw = base64.b64decode(data_b64.encode("ascii"), validate=True)
-                except Exception:
-                    _json_response(self, 400, {"error": "invalid base64"})
-                    return
-                try:
-                    out_path = _stage_uploaded_file(session_id, filename, raw)
-                except ValueError as e:
-                    status = 413 if str(e).startswith("file too large") else 400
-                    _json_response(self, status, {"error": str(e)})
-                    return
-
-                try:
-                    inject_text = _attachment_inject_text(attachment_index, out_path)
-                except ValueError as e:
-                    _json_response(self, 400, {"error": str(e)})
-                    return
-
-                # Bracketed paste: inject the staged attachment line into the active broker input.
-                seq = f"\x1b[200~{inject_text}\x1b[201~"
-                try:
-                    resp = MANAGER.inject_keys(session_id, seq)
-                except KeyError:
-                    _json_response(self, 404, {"error": "unknown session"})
-                    return
-                except ValueError as e:
-                    _json_response(self, 502, {"error": str(e)})
-                    return
-                _json_response(
-                    self,
-                    200,
-                    {
-                        "ok": True,
-                        "path": str(out_path),
-                        "inject_text": inject_text,
-                        "broker": resp,
-                    },
-                )
-                return
-
-            if path == "/api/hooks/notify":
-                # Optional integration point. Current design does not rely on this.
-                _read_body(self)
-                _json_response(self, 200, {"ignored": True})
-                return
-
             self.send_error(404)
         except KeyError:
             _json_response(self, 404, {"error": "unknown session"})
